fix(substrate): count only active memories in synthesis recap; unify memory plane to default scope; harden hex key decode
- trigger_synthesis: build_recap now receives the count of active
  (non-archived/deleted) memories instead of memories.len(), so the
  'Working set' headline no longer inflates after decay sweeps archive
  candidates. Add regression tests for the active-count and all-archived
  cases.
- list_memories / trigger_synthesis: read the single-user memory plane
  from the default scope to match the pin/unpin/forget/decay/extract
  paths, removing the latent read/write scope asymmetry (a custom scope
  could previously list ids the mutation paths could never find). The
  scope parameter is retained for N-API signature stability and
  documented as reserved; concept_graph_json remains scope-parameterized.
- decode_db_key: index over raw bytes via a hex_nibble helper so a
  malformed key with multi-byte UTF-8 can never panic on a char-boundary
  slice; non-hex input is reported as InvalidKeyHex.

+++ crates/tessera_substrate/tests/integration.rs
@@ -203,7 +203,7 @@ fn fresh_memories_survive_a_decay_sweep() {
         .unwrap();
 
     let report = manager.run_decay_sweep().expect("decay sweep");
-    assert_eq!(report.scored as u32, count, "every memory is rescored");
+    assert_eq!(report.scored, count, "every memory is rescored");
     assert_eq!(
         report.candidates_archived, 0,
         "freshly-extracted memories must not be archived immediately"
